+++ QUANTAXIS/QAData/financial_mean.py
@@ -184,7 +184,7 @@ financial_dict = {
     '157现金等价物的期初余额': 'cashEquivalentsBeginningBal',
     '158现金及现金等价物净增加额': 'netIncreaseOfCashAndCashEquivalents',
     # 5. 偿债能力分析
-    '159流动比率': 'liquidityRatio',  # 流动资产/流动负债
+    '159流动比率': 'currentRatio',  # 流动资产/流动负债
     '160速动比率': 'acidTestRatio',  # (流动资产-存货）/流动负债
     '161现金比率(%)': 'cashRatio',  # (货币资金+有价证券)÷流动负债
     '162利息保障倍数': 'interestCoverageRatio',  # (利润总额+财务费用（仅指利息费用部份）)/利息费用
@@ -242,7 +242,7 @@ financial_dict = {
     '209EBITDA/营业总收入(%)': 'EBITDA/GrossRevenueRate',
     # 9. 资本结构分析
     '210资产负债率(%)': 'assetsLiabilitiesRatio',
-    '211流动资产比率': 'liquidityRatio',
+    '211流动资产比率': 'currentAssetsRatio',# 期末的流动资产除以所有者权益
     '212货币资金比率': 'monetaryFundRatio',
     '213存货比率': 'inventoryRatio',
     '214固定资产比率': 'fixedAssetsRatio',